Split up ComponentFactory for easier use.
